Fixed some warnings   - Removed redundant variable assignment   - Replaced implicit append with explicit   - Handled possible NullPointerException   - Removed pre-sized `toArray()` call
